index: instead of comparing ourselves Flash, talk about all of the nice features that we provide
We're looking to entice people who aren't necessarily familiar with Flash's feature set. Even if someone knows what Flash could do, this is more straightforward.

The old section specifically linked to the tutorial about using SWF assets from Animate. I'm not sure it makes sense to call that out here, but it is on the features page.